+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 
 * `#81 <https://github.com/xitrum-framework/xitrum/issues/81>`_
   Reload classes in development mode
+* `#398 <https://github.com/xitrum-framework/xitrum/issues/398>`_
+  Recollect routes in development mode
 * `#393 <https://github.com/xitrum-framework/xitrum/issues/393>`_
   For development mode, set Netty's ResourceLeakDetector level to ADVANCED
 * `#395 <https://github.com/xitrum-framework/xitrum/issues/395>`_